introduced a new variable and changed the data type

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -11,7 +11,7 @@
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 {
 	void *new_ptr;
-	size_t copy_size;
+	unsigned int copy_size;
 
 	if (new_size == 0)
 	{
@@ -34,6 +34,7 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 	{
 		return (NULL);
 	}
+
 	copy_size = old_size < new_size ? old_size : new_size;
 	memcpy(new_ptr, ptr, copy_size);
 	free(ptr);
